Place actions and the reducer into their own dirs
We wanted actions, action creators and the reducer all in the same file,
as proposed by
[Redux Ducks](https://github.com/erikras/ducks-modular-redux). However,
this results in an apparently circular import during test runs that
leaves the linodeTypes module undefined upon use in some cases. This
leaves the api reducers undefined when combineReducers is eventually
called, and thus fails the test. The offending import is that of
'./fetch', i.e.  '~/api/fetch'. As mentioned this import appears to
result in a cycle, because it leaves the linodeTypes module undefined
upon use for certain tests and not others depending on what they import.
However it is not clear which modules actually compose this cycle. If
you would like to investigate, check out commit 0113d73544fe36 and see
if you can get the tests to pass.

Thefore, the actions (and base action creators) were placed into one
directory, the reducers were placed into anther directory, and the
primary actionCreator thunks were left in a 'linodeTypes' module in the
main 'api' directory. In this fashion, we can stil import the thunks
like `import * as linodeTypes from '~/api/linodeTypes';` and the tests
pass.
